Refactoring bulk table formula work a bit.
